refactor: move manifest.rs into its own module
