+++ crates/dojo-world/src/manifest/manifest_test.rs
