+++ crates/dojo-world/src/manifest/types.rs
@@ -1,0 +1,182 @@
+use camino::Utf8PathBuf;
+use serde::{Deserialize, Serialize};
+use serde_with::serde_as;
+use smol_str::SmolStr;
+use starknet::core::serde::unsigned_field_element::UfeHex;
+use starknet_crypto::FieldElement;
+
+// Collection of different types of `Manifest`'s which are used by dojo compiler/sozo
+// For example:
+//   - `BaseManifest` is generated by the compiler and wrote to `manifests/base` folder of project
+//   - `DeploymentManifest` is generated by sozo which represents the future onchain state after a successful migration
+//   - `OverlayManifest` is used by sozo to override values of specific manifest of `BaseManifest` thats generated by compiler
+
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
+pub struct BaseManifest {
+    pub world: Manifest<Class>,
+    pub base: Manifest<Class>,
+    pub contracts: Vec<Manifest<DojoContract>>,
+    pub models: Vec<Manifest<DojoModel>>,
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
+pub struct DeploymentManifest {
+    pub world: Manifest<Contract>,
+    pub base: Manifest<Class>,
+    pub contracts: Vec<Manifest<DojoContract>>,
+    pub models: Vec<Manifest<DojoModel>>,
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
+pub struct OverlayManifest {
+    pub contracts: Vec<OverlayDojoContract>,
+}
+
+#[derive(Clone, Serialize, Deserialize, PartialEq, Debug)]
+pub struct Manifest<T>
+where
+    T: ManifestMethods,
+{
+    #[serde(flatten)]
+    pub inner: T,
+    pub name: SmolStr,
+}
+
+// Utility methods thats needs to be implemented by manifest types
+pub trait ManifestMethods {
+    type OverlayType;
+    fn abi(&self) -> Option<&Utf8PathBuf>;
+    fn set_abi(&mut self, abi: Option<Utf8PathBuf>);
+    fn class_hash(&self) -> &FieldElement;
+    fn set_class_hash(&mut self, class_hash: FieldElement);
+
+    /// This method is called when during compilation base manifest file already exists.
+    /// Manifest generated during compilation won't contains properties manually updated by users
+    /// (like calldata) so this method should override those fields
+    fn merge(&mut self, old: Self::OverlayType);
+}
+
+impl<T> Manifest<T>
+where
+    T: ManifestMethods,
+{
+    pub fn new(inner: T, name: SmolStr) -> Self {
+        Self { inner, name }
+    }
+}
+
+#[serde_as]
+#[derive(Clone, Default, Debug, Serialize, Deserialize, PartialEq)]
+#[serde(tag = "kind")]
+pub struct DojoContract {
+    #[serde_as(as = "Option<UfeHex>")]
+    pub address: Option<FieldElement>,
+    #[serde_as(as = "UfeHex")]
+    pub class_hash: FieldElement,
+    pub abi: Option<Utf8PathBuf>,
+    pub reads: Vec<String>,
+    pub writes: Vec<String>,
+    pub computed: Vec<ComputedValueEntrypoint>,
+}
+
+/// Represents a declaration of a model.
+#[serde_as]
+#[derive(Clone, Default, Debug, Serialize, Deserialize, PartialEq)]
+#[serde(tag = "kind")]
+pub struct DojoModel {
+    pub members: Vec<Member>,
+    #[serde_as(as = "UfeHex")]
+    pub class_hash: FieldElement,
+    pub abi: Option<Utf8PathBuf>,
+}
+
+#[serde_as]
+#[derive(Clone, Default, Debug, Serialize, Deserialize, PartialEq)]
+#[serde(tag = "kind")]
+pub struct Contract {
+    #[serde_as(as = "UfeHex")]
+    pub class_hash: FieldElement,
+    pub abi: Option<Utf8PathBuf>,
+    #[serde_as(as = "Option<UfeHex>")]
+    pub address: Option<FieldElement>,
+    #[serde_as(as = "Option<UfeHex>")]
+    pub transaction_hash: Option<FieldElement>,
+    pub block_number: Option<u64>,
+    // used by World contract
+    pub seed: Option<String>,
+}
+
+#[serde_as]
+#[derive(Clone, Default, Debug, Serialize, Deserialize, PartialEq)]
+#[serde(tag = "kind")]
+pub struct Class {
+    #[serde_as(as = "UfeHex")]
+    pub class_hash: FieldElement,
+    pub abi: Option<Utf8PathBuf>,
+}
+
+#[serde_as]
+#[derive(Clone, Default, Debug, Serialize, Deserialize, PartialEq)]
+
+pub struct OverlayDojoContract {
+    pub name: SmolStr,
+    pub reads: Option<Vec<String>>,
+    pub writes: Option<Vec<String>>,
+}
+
+#[serde_as]
+#[derive(Clone, Default, Debug, Serialize, Deserialize, PartialEq)]
+pub struct OverlayDojoModel {}
+
+#[serde_as]
+#[derive(Clone, Default, Debug, Serialize, Deserialize, PartialEq)]
+pub struct OverlayContract {}
+
+#[serde_as]
+#[derive(Clone, Default, Debug, Serialize, Deserialize, PartialEq)]
+pub struct OverlayClass {}
+
+// Types used by manifest
+
+/// Represents a model member.
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
+pub struct Member {
+    /// Name of the member.
+    pub name: String,
+    /// Type of the member.
+    #[serde(rename = "type")]
+    pub ty: String,
+    pub key: bool,
+}
+
+#[serde_as]
+#[derive(Clone, Default, Debug, Serialize, Deserialize, PartialEq)]
+pub struct ComputedValueEntrypoint {
+    // Name of the contract containing the entrypoint
+    pub contract: SmolStr,
+    // Name of entrypoint to get computed value
+    pub entrypoint: SmolStr,
+    // Component to compute for
+    pub model: Option<String>,
+}
+
+impl From<dojo_types::schema::Member> for Member {
+    fn from(m: dojo_types::schema::Member) -> Self {
+        Self { name: m.name, ty: m.ty.name(), key: m.key }
+    }
+}
+
+/// System input ABI.
+#[derive(Clone, Default, Debug, Serialize, Deserialize, PartialEq)]
+pub struct Input {
+    pub name: String,
+    #[serde(rename = "type")]
+    pub ty: String,
+}
+
+/// System Output ABI.
+#[derive(Clone, Default, Debug, Serialize, Deserialize, PartialEq)]
+pub struct Output {
+    #[serde(rename = "type")]
+    pub ty: String,
+}
